requirements.txt
django==1.7.11
requests
chardet
patool
python-ldap
